Rename stories dir to the stories_tree to prevent prepublish script ignore flags collision
