+++ lib/ui/src/modules/ui/components/left_panel/index.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import pick from 'lodash.pick';
 import Header from './header';
-import Stories from './stories';
+import Stories from './stories_tree';
 import TextFilter from './text_filter';
 
 const scrollStyle = {
