+++ lib/ui/src/modules/ui/components/left_panel/index.test.js
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 import LeftPanel from './index';
 import Header from './header';
 import TextFilter from './text_filter';
-import Stories from './stories';
+import Stories from './stories_tree';
 import { createHierarchy } from '../../libs/hierarchy';
 
 describe('manager.ui.components.left_panel.index', () => {
