+++ lib/ui/src/modules/ui/components/left_panel/stories_tree/index.js
