+++ lib/ui/src/modules/ui/components/left_panel/stories_tree/tree_style.js
